+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename "i-card" to "Illini ID" in Settings [#3588](https://github.com/rokwire/illinois-app/issues/3588).
 ### Fixed
 - Show Twitter in favorites [#3590](https://github.com/rokwire/illinois-app/issues/3590).
+### Added
+- Added language selector in Settings panel [#3587](https://github.com/rokwire/illinois-app/issues/3587).
+
 
 ## [5.0.49] - 2023-08-17
 ### Changed